DRASTIC SIMPLIFICATIONS / MEMORY OPTIMIZATIONS
I am trying to improve the slurm resource estimation in my code.
i am trying to use less ram -> dont store states -> store eops instead

also instead of storing all phi1,2 combinations, accumulate them on the fly to save memory

TODO implement one version which is fast/memory efficient "e_field_1d" and one that is good for testing

outsourced phase cycling to a separate module

implemented 0d, 1,d 2d case of calculation by changing t_det_max to t_det
also t_coh: will be the maximal coherence time so

TODO test tomorrow!
-> first one small test on hpc to see if ram simplifications work

+++ packages/qspectro2d/src/qspectro2d/spectroscopy/polarization.py
@@ -25,9 +25,9 @@ def complex_polarization(
         return _single_qobj__complex_pol(dipole_op, state)
     if isinstance(state, list):
         if len(state) == 0:
-            return np.array([], dtype=np.complex64)
+            return np.array([], dtype=np.complex128)
         return np.array(
-            [_single_qobj__complex_pol(dipole_op, s) for s in state], dtype=np.complex64
+            [_single_qobj__complex_pol(dipole_op, s) for s in state], dtype=np.complex128
         )
     raise TypeError(f"State must be Qobj or list[Qobj], got {type(state)}")
